Split out OptionParser
